data_fit2 {mbImpute}	R Documentation

data_fit2

Description

data_fit2 is a function that aims to perform one imputation in a group of biological samples, it's a subfunction of mbImpute and intend not to be used by the user.

Usage

data_fit2(y_sim, metadata, D, k, parallel = F, ncores = 1)

Arguments

y_sim
A matrix of dimension n * m corresponding to the normalized OTU table, with n denoting the number of subjects(patients), and m denoting the number of taxa.

metadata
A matrix of dimension n * p corresponding to the meta data matrix, with n denotes the number of subjects(patients), and p denotes the number of covariates (age/BMI/...).

D
A matrix of dimension m * m corresponding to the taxa distance matrix.

k
A scalar corresponding to the number of nearest taxa in a phylogenetic tree we will use to impute a missing value. Theoretically, the larger k, the more accurate our imputation will be (required that k <= m).

parallel
A boolean indicating whether to use parallel or not. Default is FALSE.

ncores
A scalar corresponding to the number of cores to use. It is only used when parallel = TRUE. Default is 1.

Value

y_imp: the imputed matrix

[Package mbImpute version 0.1.0 Index]