distance_mat_gen {mbImpute}	R Documentation

distance_mat_gen

Description

This function aims to generate the distance matrix of taxa based on their path lengths in the phylogenetic tree.

Usage

distance_mat_gen(edges, n_taxa, tree_height = 50)

Arguments

edges
A matrix of dimension N * 2 corresponding to the edge set of the phylogenetic tree (similar to the edge set for a graph).

n_taxa
A scalar corresponding to number of taxa in the dataset.

tree_height
A scalar corresponding to the height of the phylogenetic tree. Any number larger than the height of the phylogenetic tree will also work. The initial value is set to be 50, which is usually enough as for a complete binary tree, height of 50 corresponds to 2^50 nodes.

Value

A matrix of dimension n_taxa * n_taxa corresponding to the matrix D in mbImpute function.

[Package mbImpute version 0.1.0 Index]