mbImpute {mbImpute}	R Documentation

mbImpute: an accurate and robust imputation method for microbiome data

Description

mbImpute aims to impute likely false zero counts or low counts for a normalized OTU table of dimension n * m. Where n is the number of biological samples and m is the number of taxa. In order to achieve our goal, we will borrow information from covariate matrix of dimension n * p and taxa distance matrix of dimension m * m. The final result of the mbImpute function will ouput an imputed matrix of dimension n * m, which is exactly the same dimension as the input.

Usage

mbImpute(
  condition = NULL,
  otu_tab = NULL,
  metadata = NULL,
  D = NULL,
  k = 5,
  parallel = FALSE,
  ncores = 1,
  unnormalized = T
)

Arguments

condition
A vector of length n indicating which group (e.g. treatment/control, CRC/control/Adenoma) the biology samples belong to.

otu_tab
A matrix of dimension n * m corresponding to OTU count table.

metadata
A matrix of dimension n * p corresponding to the covariate matrix/meta data comes with the biological sample. Please note that the conditions/groups of the samples should be excluded from the meta data.

D
A matrix of dimension m * m corresponding to the distance matrix of taxa generated from an edge set. If an edge set is provided, you can use distance_mat function in our package to generate the distance matrix D.

k
A scalar corresponding to the number of nearest taxa in a phylogenetic tree we will use to impute a missing value. Theoretically, the larger k, the more accurate our imputation will be (required that k <= m).

parallel
A boolean indicating whether to use parallel or not. Default is FALSE.

ncores
A scalar corresponding to the number of cores to use. It is only used when parallel = TRUE. Default is 1.

unnormalized
A boolean indicating whether the data has been normalized. Details can be found in the Vignette.

Value

A list three imputed OTU matrices. imp_count_mat_lognorm: imputed normalized and log transformed matrix. imp_count_mat_norm : imputed normalized count matrix with library size of each sample equal to 10^6. imp_count_mat_origlibsize: imputed countmatrix at the original library size. We recommend to use the first one imp_count_mat_lognorm. So that the log scaled counts follows approximately Normal for each taxon across samples.

[Package mbImpute version 0.1.0 Index]